% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_split.R
\name{eval_split}
\alias{eval_split}
\title{eval_split}
\usage{
eval_split(split_df, method = "ls")
}
\arguments{
\item{split_df}{A matrix with two columns coming out of the "split" function}

\item{method}{A string giving the evaluation method (default is "ls")
--TO BE EXTENDED}
}
\value{
A named vector with the preferred tree node after evaluation.
The value gives the outcome in the node, the name is "0" or "1" depending on
which of the two nodes was selected.
}
\description{
Helper function to evaluate the current split based on a given evaluation
method (the default is least squares, "ls")
}
\examples{
pref_node = eval_split(split(pred1, out), method = "ls")
}
