% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{split}
\usage{
split(pred, out)
}
\arguments{
\item{pred}{A vector of zeroes and ones}

\item{out}{A vector of numeric or intreger values}
}
\value{
A matrix with two columns, the first being the values of the outcome
corresponding to the 0 value of the predictor, the second to 1.
}
\description{
Helper function to perform a split on a categorical predictor
}
\examples{
split1 = split(pred1, out)
}
