% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{tree}
\usage{
tree(df)
}
\arguments{
\item{df}{}
}
\value{
--TBD
}
\description{
Builds the decision tree. Uses helper functions "split" and "eval_split" to
perform and evaluate splits.
}
\examples{
--TBD
}
