% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{simdata}
\usage{
simdata(n_obs, n_vars, predictors = "factor", output = "cont")
}
\arguments{
\item{output}{}
}
\value{
a dataframe with multiple predictors and a single outcome variable
}
\description{
simulate some data to test the tree on
}
\examples{
dat = simdata(500, 5)
head(dat)
}
